Use new WinRT APIs for SpecialFolders (#22664)
* Use new WinRT APIs for SpecialFolders

* Update tests

* Update test package

Fix test verification

* Clean up project files

* Add back test skip condition

* Add comments and fallback Windows dir.
