+++ external/test-runtime/XUnit.Runtime.depproj
@@ -94,7 +94,7 @@
   </ItemGroup>
 
   <PropertyGroup Condition="'$(TargetGroup)' == 'uap'">
-    <UAPToolsPackageVersion>1.0.14</UAPToolsPackageVersion>
+    <UAPToolsPackageVersion>1.0.15</UAPToolsPackageVersion>
   </PropertyGroup>
 
   <ItemGroup Condition="'$(TargetGroup)' == 'uap'">
